requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
django-multiselectfield==0.1.12
djangorestframework==3.13.1
idna==3.3
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
